+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 * Fix routing loop caused by incorrect Custom Fields routing. Refs UIU-1594.
 * Settings > Users > Create/Edit Permission set. Refs UIU-1587.
 * Add `Closed loan` translation. Part of UIU-1603.
+* Pass current `servicePointId` to `declare-item-lost`. Part of UIU-1203.
 
 ## [3.0.0](https://github.com/folio-org/ui-users/tree/v3.0.0) (2020-03-17)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v2.26.0...v3.0.0)